requirements.txt
urllib3
azure-identity